+++ ext/github/RepositoryKey.py
@@ -41,16 +41,16 @@ class RepositoryKey(github.GithubObject.CompletableGithubObject):
     This class represents RepositoryKeys. The reference can be found here http://developer.github.com/v3/repos/keys/
     """
 
-    def __init__(self, requester, headers, attributes, completed, repoUrl):
-        github.GithubObject.CompletableGithubObject.__init__(self, requester, headers, attributes, completed)
-        self.__repoUrl = repoUrl
-
     def __repr__(self):
-        return self.get__repr__({"id": self._id.value})
+        return self.get__repr__({"id": self._id.value, "title": self._title.value})
 
     @property
-    def __customUrl(self):
-        return self.__repoUrl + "/keys/" + str(self.id)
+    def created_at(self):
+        """
+        :type: datetime.datetime
+        """
+        self._completeIfNotSet(self._created_at)
+        return self._created_at.value
 
     @property
     def id(self):
@@ -107,31 +107,11 @@ class RepositoryKey(github.GithubObject.CompletableGithubObject):
         """
         headers, data = self._requester.requestJsonAndCheck(
             "DELETE",
-            self.__customUrl
+            self.url
         )
-
-    def edit(self, title=github.GithubObject.NotSet, key=github.GithubObject.NotSet):
-        """
-        :calls: `PATCH /repos/:owner/:repo/keys/:id <http://developer.github.com/v3/repos/keys>`_
-        :param title: string
-        :param key: string
-        :rtype: None
-        """
-        assert title is github.GithubObject.NotSet or isinstance(title, (str, unicode)), title
-        assert key is github.GithubObject.NotSet or isinstance(key, (str, unicode)), key
-        post_parameters = dict()
-        if title is not github.GithubObject.NotSet:
-            post_parameters["title"] = title
-        if key is not github.GithubObject.NotSet:
-            post_parameters["key"] = key
-        headers, data = self._requester.requestJsonAndCheck(
-            "PATCH",
-            self.__customUrl,
-            input=post_parameters
-        )
-        self._useAttributes(data)
 
     def _initAttributes(self):
+        self._created_at = github.GithubObject.NotSet
         self._id = github.GithubObject.NotSet
         self._key = github.GithubObject.NotSet
         self._title = github.GithubObject.NotSet
@@ -140,6 +120,8 @@ class RepositoryKey(github.GithubObject.CompletableGithubObject):
         self._read_only = github.GithubObject.NotSet
 
     def _useAttributes(self, attributes):
+        if "created_at" in attributes:  # pragma no branch
+            self._created_at = self._makeDatetimeAttribute(attributes["created_at"])
         if "id" in attributes:  # pragma no branch
             self._id = self._makeIntAttribute(attributes["id"])
         if "key" in attributes:  # pragma no branch