+++ themes-default/slim/static/js/router.js
@@ -3,6 +3,8 @@ const { routes, httpVueLoader } = window;
 
 if (!window.router) {
     const configComponent = httpVueLoader('js/templates/config.vue');
+    const addShowsComponent = httpVueLoader('js/templates/add-shows.vue');
+    const addRecommendedComponent = httpVueLoader('js/templates/add-recommended.vue');
     const notFoundComponent = httpVueLoader('js/templates/http/404.vue');
 
     const router = new VueRouter({
@@ -16,6 +18,22 @@ if (!window.router) {
                 header: 'Medusa Configuration'
             },
             component: configComponent
+        }, {
+            path: '/addShows',
+            name: 'addShows',
+            meta: {
+                title: 'Add Shows',
+                header: 'Add Shows'
+            },
+            component: addShowsComponent
+        }, {
+            path: '/addRecommended',
+            name: 'addRecommended',
+            meta: {
+                title: 'Add Recommended Shows',
+                header: 'Add Recommended Shows'
+            },
+            component: addRecommendedComponent
         }, {
             path: '/not-found',
             name: 'not-found',