+++ CHANGELOG.md
@@ -2,14 +2,21 @@
 
 **New Features**
 
-- Simple message describing the new feature, and a link to the pull request.
+- Hot-swap themes: No need to restart the Medusa after changing the theme ([#4271](https://github.com/pymedusa/Medusa/pull/4271))
+- _Simple message describing the new feature, and a link to the pull request._
 
 **Improvements**
 
-- Simple message describing the improvement, and a link to the pull request.
+- Moved the following routes to use `VueRouter` + `http-vue-loader`:
+  - `/config` - Help & Info ([#4374](https://github.com/pymedusa/Medusa/pull/4374))
+  - `/addShows` - Add Shows ([#4564](https://github.com/pymedusa/Medusa/pull/4564))
+  - `/addRecommended` - Add Recommended Shows ([#4564](https://github.com/pymedusa/Medusa/pull/4564))
+- Removed the old `/ui` route ([#4565](https://github.com/pymedusa/Medusa/pull/4565))
+- _Simple message describing the improvement, and a link to the pull request._
 
 **Fixes**
 
-- Simple message describing the fix, and a link to the pull request.
+- Fixed Emby test notification ([#4622](https://github.com/pymedusa/Medusa/pull/4622))
+- _Simple message describing the fix, and a link to the pull request._
 
 ### [**Previous versions**](https://github.com/pymedusa/medusa.github.io/blob/master/news/CHANGES.md)